@@ -1,47 +1,42 @@
-const PubSub = require('../helpers/pub_sub.js');
+/* eslint-disable class-methods-use-this */
+import { subscribe, publish } from '../helpers/pubSub.js';
 
-const Board = function () {
-};
+export default class Board {
+  constructor() {}
 
-Board.prototype.bindEvents = function () {
+  bindEvents() {
+    subscribe('Game:player-choose-move', (evt) => {
+      this.highlightSquares(evt.detail[0], evt.detail[1]);
+    });
+  }
 
-  PubSub.subscribe('Game:player-choose-move', (evt) => {
-    this.highlightSquares(evt.detail[0], evt.detail[1]);
+  setBoardPieces(numberOfPlayers) {
+    document.querySelector('#p1-piece').style.display = 'block';
+    document.querySelector('#p2-piece').style.display = 'block';
+    if (numberOfPlayers > 3) document.querySelector('#p4-piece').style.display = 'block';
+    if (numberOfPlayers > 2) document.querySelector('#p3-piece').style.display = 'block';
+  }
 
-  });
+  highlightSquares(squares ,player) {
+    squares.forEach((moveOption) => {
+      const highlightedSquare = document.querySelector(`#${moveOption}`);
+      highlightedSquare.classList.add('highlighted-move');
+      highlightedSquare.style.border = '3px white';
+      highlightedSquare.style.borderStyle = 'dashed solid';
+      highlightedSquare.style.color = 'white';
+      highlightedSquare.onclick = function () {
+        this.clearHighlightedSquares();
+        publish('Board:player-move', this.id);
+      };
+    });
+  }
 
-};
-
-Board.prototype.setBoardPieces = function (number_of_players) {
-  document.querySelector('#p1-piece').style.display = 'block';
-  document.querySelector('#p2-piece').style.display = 'block';
-  if (number_of_players>3) document.querySelector('#p4-piece').style.display = 'block';
-  if (number_of_players>2) document.querySelector('#p3-piece').style.display = 'block';
-};
-
-Board.prototype.highlightSquares = function (squares, player) {
-  for (move_option of squares) {
-    const highlightedSquare = document.querySelector(`#${move_option}`);
-    highlightedSquare.classList.add('highlighted-move');
-    highlightedSquare.style.border = "3px white";
-    highlightedSquare.style.borderStyle = "dashed solid";
-    highlightedSquare.style.color = "white";
-    highlightedSquare.onclick = function () {
-      clearHighlightedSquares();
-      PubSub.publish('Board:player-move', this.id);
-    };
-  };
-};
-
-
-clearHighlightedSquares = function () {
-  for (square of document.querySelectorAll('.highlighted-move')) {
-    square.classList.remove('highlighted-move');
-    square.style.border = "none";
-    square.style.color = "black";
-    square.onclick = null;
-  };
-};
-
-
-module.exports = Board;
+  clearHighlightedSquares() {
+    document.querySelectorAll('.highlighted-move').forEach((square) => {
+      square.classList.remove('highlighted-move');
+      square.style.border = 'none';
+      square.style.color = 'black';
+      square.onclick = null;
+    });
+  }
+}
