@@ -1,24 +1,40 @@
-const express = require('express');
-const app = express();
-const path = require('path');
-const parser = require('body-parser');
-const MongoClient = require('mongodb').MongoClient;
-const createRouter = require('./helpers/create_router.js');
+import url from 'url';
+import express from 'express';
+import { join } from 'path';
+import bodyParser from 'body-parser';
+import mongodb from 'mongodb';
+import createRouter from './helpers/createRouter.js';
 
-const publicPath = path.join(__dirname, '../client/public');
+const { json } = bodyParser;
+const { MongoClient } = mongodb;
+const app = express();
+
+const publicPath = join(url.fileURLToPath(new URL('.', import.meta.url)), '../client/public');
 app.use(express.static(publicPath));
 
-app.use(parser.json());
+app.use(json());
 
-MongoClient.connect('mongodb://heroku_sdpqx20n:vu751c1jh8j0ciigp6to9k7093@ds151840.mlab.com:51840/heroku_sdpqx20n')
-  .then((client) => {
-    const db = client.db('heroku_sdpqx20n');
+const uri = 'mongodb://localhost:27017';
+
+const client = MongoClient(uri);
+
+async function run() {
+  try {
+    await client.connect();
+
+    const db = await client.db('pieInTheSky');
+    db.command({ ping: 1 });
+    console.log('Connected successfully to db');
     const gamesCollection = db.collection('games');
     const gamesRouter = createRouter(gamesCollection);
     app.use('/api/games', gamesRouter);
-  })
-  .catch(console.err);
+  } finally {
+    await client.close();
+  }
+}
 
-app.listen(process.env.PORT || 8080, function () {
-  console.log(`Listening on port ${ this.address().port}, ${ app.settings.env }`);
+await run().catch(console.dir);
+
+app.listen(process.env.PORT || 3000, function () {
+  console.log(`Listening on port ${this.address().port}, ${app.settings.env}`);
 });
